# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlatRadar Backend API",
    "version": "1.6.0",
    "description": "Machine-readable contract for FlatRadar mobile clients. All JSON endpoints use a stable { ok, data | error } envelope."
  },
...
  "servers": [
    {
      "url": "https://flatradar.app/api/v1",
      "description": "Production"
    },
    {
      "url": "http://127.0.0.1:8088/api/v1",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/notifications": {
      "get": {
        "tags": ["Notifications"],
        "operationId": "listNotifications",
        "summary": "List visible notifications.",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/LimitQuery"},
          {"$ref": "#/components/parameters/OffsetQuery"}
        ],
        "responses": {
          "200": {"$ref": "#/components/responses/NotificationsOk"},
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      }
    },
    "/notifications/read": {
      "post": {
        "tags": ["Notifications"],
        "operationId": "markNotificationsRead",
        "summary": "Mark selected or all notifications as read.",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/NotificationsReadRequest"}
            }
          }
        },
        "responses": {
          "200": {"$ref": "#/components/responses/NotificationsReadOk"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
...
      "LimitQuery": {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 500, "default": 100}},
      "OffsetQuery": {"name": "offset", "in": "query", "schema": {"type": "integer", "minimum": 0, "default": 0}},
...
    },
    "responses": {
      "BadRequest": {"description": "Validation error.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Unauthorized": {"description": "Missing, invalid, expired, or disabled bearer identity.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
...
      "NotificationsOk": {"description": "Notification page.", "content": {"application/json": {"schema": {"allOf": [{"$ref": "#/components/schemas/SuccessEnvelope"}, {"type": "object", "properties": {"data": {"$ref": "#/components/schemas/NotificationsPage"}}}]}}}},
      "NotificationsReadOk": {"description": "Notifications marked.", "content": {"application/json": {"schema": {"allOf": [{"$ref": "#/components/schemas/SuccessEnvelope"}, {"type": "object", "properties": {"data": {"type": "object", "required": ["marked"], "properties": {"marked": {"type": "boolean"}}}}}]}}}},
...
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": ["ok", "data"],
        "properties": {
          "ok": {"type": "boolean", "const": true},
          "data": {}
        },
        "additionalProperties": false
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["ok", "error"],
        "properties": {
          "ok": {"type": "boolean", "const": false},
          "error": {"$ref": "#/components/schemas/ApiError"}
        },
        "additionalProperties": false
      },
      "ApiError": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {"$ref": "#/components/schemas/ApiErrorCode"},
          "message": {"type": "string"}
        },
        "additionalProperties": false
      },
      "ApiErrorCode": {
        "type": "string",
        "enum": ["unauthorized", "forbidden", "not_found", "validation", "conflict", "rate_limited", "server_error"]
      },
...
      "Notification": {
        "type": "object",
        "required": ["id", "created_at", "type", "title", "body", "read"],
        "properties": {
          "id": {"type": "integer"},
          "created_at": {"type": "string"},
          "type": {"type": "string"},
          "title": {"type": "string"},
          "body": {"type": "string"},
          "url": {"type": "string"},
          "listing_id": {"type": ["string", "null"]},
          "read": {"type": "integer", "enum": [0, 1]},
          "user_id": {"type": ["string", "null"]}
        },
        "additionalProperties": true
      },
      "NotificationsPage": {
        "type": "object",
        "required": ["items", "total", "unread", "limit", "offset"],
        "properties": {
          "items": {"type": "array", "items": {"$ref": "#/components/schemas/Notification"}},
          "total": {"type": "integer", "minimum": 0},
          "unread": {"type": "integer", "minimum": 0},
          "limit": {"type": "integer", "minimum": 1},
          "offset": {"type": "integer", "minimum": 0}
        },
        "additionalProperties": true
      },
      "NotificationsReadRequest": {
        "type": "object",
        "properties": {
          "ids": {"type": "array", "items": {"type": "integer"}}
        },
        "additionalProperties": false
      },
...
    }
  }
}